WIP rework of pipelining for 32-bit intermediate buffer * removes now unused 8 bpp -> 8 bpp blitters * adds auxiliary BlitMosaic() and Blit32_32() that work with 32-bit source buffers and allow optional blending * reworks drawing of background layers, moving repeated copy'n paste blocks to reusable functions * places variable declarations where used instead of cluttering start of function (old C89 style) * mosaic, affine and pixel mapping use 32-bit intermediary buffer and new auxiliary blitters instead of 8-bit ones.

+++ src/Layer.h
@@ -51,8 +51,8 @@ typedef struct Layer
 	bool dirty;					/* requires UpdateLayer() before draw */
 	
 	/* */
-	int				hstart;		/* offset de inicio horizontal */
-	int				vstart;		/* offset de inicio vertical */
+	int				hstart;		/* horizontal start offset */
+	int				vstart;		/* vertical start offset*/
 
 	/* clip */
 	struct
@@ -64,8 +64,8 @@ typedef struct Layer
 	/* mosaic */
 	struct
 	{
-		int w,h;			/* tamaño del pixel */
-		uint8_t* buffer;	/* linea temporal */
+		int w, h;			/* virtual pixel size */
+		uint32_t* buffer;	/* line buffer */
 	}
 	mosaic;
 }
